+++ src/market-making/leaves/SystemDebt.sol
@@ -16,8 +16,8 @@ library SystemDebt {
 
     // TODO: pack storage slots
     struct Data {
-        int256 totalUnsettledDebtUsd;
-        int256 totalSettledDebtUsd;
+        int128 totalUnsettledDebtUsd;
+        int128 totalSettledDebtUsd;
         MinHeapLib.Heap vaultsDebtSettlementPriorityQueue;
     }
